Rename adapters to adapter
