+++ src/balancing/mod.rs
@@ -4,7 +4,7 @@ mod middleware;
 pub use self::{
     frontend::Server,
     middleware::{
-        adapters::{chain::chain, proj::snd},
+        adapter::{chain::chain, proj::snd},
         client::connect,
         route::route,
         selector::cyclical::RoundRobin,
